♻️ Split editor form components
